A new unit test.
